```
% go-perftuner benchstat 1.txt 2.txt 
args: [1.txt 2.txt]name    old time/op    new time/op    delta
PAE-10     245ns ± 4%      56ns ± 3%  -77.19%  (p=0.000 n=10+10)

name    old alloc/op   new alloc/op   delta
PAE-10      632B ± 0%      352B ± 0%  -44.30%  (p=0.000 n=10+10)

name    old allocs/op  new allocs/op  delta
PAE-10      9.00 ± 0%      1.00 ± 0%  -88.89%  (p=0.000 n=10+10)
```